Improved aimbot
Aimbot will now run 10 iterations to attempt to find the future location of the ship.

Previously, the aimbot would calculate the future position of the ship by simply taking its speed and multiplying it by the flight time of the cannonball. This has been reworked so that it will recalculate the new future position of the ship based on the old future position of the ship. It will iterate through this process 10 times.

This should make it more accurate as it narrows down the final aiming position algebraically. This could be made more efficient by doing it with calculus.

THIS HAS NOT BEEN THOROUGHLY TESTED. IT COULD BE A COMPLETELY BROKEN FEATURE THAT IS NO BETTER THAN THE OLD VERSION.